feat: Try/show backward compatibility for so101

+++ lerobot/common/motors/motors_bus.py
@@ -825,7 +825,7 @@ class MotorsBus(abc.ABC):
             elif self.motors[motor].norm_mode is MotorNormMode.DEGREES:
                 mid = (min_ + max_) / 2
                 max_res = self.model_resolution_table[self._id_to_model(id_)] - 1
-                unnormalized_values[id_] = (val * max_res / 360) + mid
+                unnormalized_values[id_] = int((val * max_res / 360) + mid)
             else:
                 raise NotImplementedError
